refactor(tamanu): move match_names from lifecycle to services
The matcher is the only pure piece of lifecycle.rs — it operates on
Expectation and has no IO. Moving it to services.rs alongside the
Expectation type means logs.rs (which doesn't pull in the
tamanu-lifecycle feature) can share the same primitive without a
cross-feature dependency.
