+++ crates/bestool/src/actions/tamanu/restart.rs
@@ -11,7 +11,7 @@ use crate::actions::{
 	tamanu::{
 		TamanuArgs,
 		lifecycle::{self, Instance},
-		services::{Criticality, ExpectedState, Expectation, Supervisor},
+		services::{self, Criticality, ExpectedState, Expectation, Supervisor},
 	},
 };
 
@@ -58,7 +58,7 @@ pub async fn run(args: RestartArgs, ctx: Context) -> Result<()> {
 
 	let (supervisor, expectations) = lifecycle::config_and_expectations(tamanu)?;
 	let names: Vec<&str> = args.names.iter().map(String::as_str).collect();
-	let matched = lifecycle::match_names(&expectations, &names)?;
+	let matched = services::match_names(&expectations, &names)?;
 	let discovered = lifecycle::discover(supervisor)?;
 	let groups = lifecycle::group_by_expectation(&matched, &discovered);
