+++ crates/bestool/src/actions/tamanu/status.rs
@@ -8,7 +8,7 @@ use crate::actions::{
 	tamanu::{
 		TamanuArgs,
 		lifecycle::{self, Instance},
-		services::{Criticality, ExpectedState, Expectation},
+		services::{self, Criticality, ExpectedState, Expectation},
 	},
 };
 
@@ -35,7 +35,7 @@ pub async fn run(args: StatusArgs, ctx: Context) -> Result<()> {
 
 	let (supervisor, expectations) = lifecycle::config_and_expectations(tamanu)?;
 	let names: Vec<&str> = args.names.iter().map(String::as_str).collect();
-	let matched = lifecycle::match_names(&expectations, &names)?;
+	let matched = services::match_names(&expectations, &names)?;
 	let discovered = lifecycle::discover(supervisor)?;
 	let groups = lifecycle::group_by_expectation(&matched, &discovered);
